backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx>=0.27.0
chromadb==0.5.3
google-generativeai==0.3.2
pypdf2==3.0.1
python-docx==1.1.0
aiofiles==23.2.1
pydantic>=2.9.0
pydantic-settings>=2.5.0
authlib==1.3.0
email-validator==2.1.0
numpy<2.0
python-telegram-bot>=20.8

